requirements.txt
fastapi
uvicorn
python-multipart
morecantile
rio-tiler
pydelatin
pymartini
quantized-mesh-encoder
aiosqlite
aiofiles
triangle
lxml
requests
requests-oauthlib
psutil
pyee>=13
jinja2
python-dotenv
